fix blockchain type in indexer, add scripts

+++ Assets/SequenceSDK/SequenceIndexer/Scripts/DataTypes/ContractInfo.cs
@@ -1,9 +1,11 @@
+using System.Numerics;
+
 namespace SequenceSharp
 {
     [System.Serializable]
     public class ContractInfo
     {
-        public int chainId;
+        public BigInteger chainId;
         public string address;
         public string name;
         public string type;
