+++ Assets/SequenceSDK/SequenceIndexer/Scripts/DataTypes/EventLog.cs
@@ -1,17 +1,19 @@
+using System.Numerics;
+
 namespace SequenceSharp
 {
     [System.Serializable]
     public class EventLog
     {
-        public int id;
+        public BigInteger id;
         public EventLogType type;
-        public int blockNumber;
+        public BigInteger blockNumber;
         public string blockHash;
         public string contractAddress;
         public ContractType contractType;
         public string txnHash;
-        public int txnIndex;
-        public int txnLogIndex;
+        public BigInteger txnIndex;
+        public BigInteger txnLogIndex;
         public EventLogDataType logDataType;
         public string ts;
         public string logData;
